refactor: fix #102
