+++ src/components/header/index.tsx
@@ -3,11 +3,13 @@ import { useRouter } from "next/router";
 import { useTranslation } from "next-i18next";
 import React from "react";
 
-import theme from "../libs/colorTheme";
+import {FloatMenu} from "@/components/floatMenu";
 
-import FloatMenu from "./floatMenu";
-import LoginBtn from "./headerModules/loginBtn";
-import Logo from "./headerModules/logo";
+import theme from "../../libs/colorTheme";
+
+import LoginBtn from "./loginBtn";
+import Logo from "./logo";
+
 
 export default React.memo(function Header() {
 	const { t } = useTranslation("components");
